Fixed errors in cast.
