+++ tests/it/compute/mod.rs
@@ -1,0 +1,1 @@
+mod cast;
