+++ tests/it/main.rs
@@ -6,3 +6,5 @@ mod ffi;
 
 mod io;
 mod test_util;
+
+mod compute;
